+++ src/modules/servicos/servicos.controller.ts
@@ -1,51 +1,7 @@
-import { Body, Controller, Param, Put } from '@nestjs/common';
-import { ApiNotFoundResponse, ApiOkResponse, ApiTags, ApiBody } from '@nestjs/swagger';
+import { Controller } from '@nestjs/common';
 import { ServicosService } from './servicos.service';
-import { UpdateSolicitacaoStatusDto } from './dto/update-solicitacao-status.dto';
 
-@ApiTags('servicos')
-@Controller()
+@Controller('servicos')
 export class ServicosController {
   constructor(private readonly servicosService: ServicosService) {}
-
-  @Put('solicitacao/:id')
-  @ApiBody({ 
-    type: UpdateSolicitacaoStatusDto,
-    description: 'Dados para atualização do status da solicitação'
-  })
-  @ApiOkResponse({ 
-    description: 'Status da solicitação atualizado com sucesso',
-    schema: {
-      type: 'object',
-      properties: {
-        message: { 
-          type: 'string', 
-          example: 'Status da solicitação atualizado com sucesso.' 
-        }
-      }
-    }
-  })
-  @ApiNotFoundResponse({ 
-    description: 'Solicitação não encontrada',
-    schema: {
-      type: 'object',
-      properties: {
-        message: { 
-          type: 'string', 
-          example: 'Solicitação com ID 999 não encontrada' 
-        },
-        error: { type: 'string', example: 'Not Found' },
-        statusCode: { type: 'number', example: 404 }
-      }
-    }
-  })
-  async updateSolicitacaoStatus(
-    @Param('id') id: string,
-    @Body() updateSolicitacaoStatusDto: UpdateSolicitacaoStatusDto,
-  ) {
-    return this.servicosService.updateSolicitacaoStatus(
-      parseInt(id, 10),
-      updateSolicitacaoStatusDto,
-    );
-  }
 }
