+++ src/modules/solicitacao/solicitacao.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { SequelizeModule } from '@nestjs/sequelize';
+import { SolicitacaoService } from './solicitacao.service';
 import { Servico } from 'src/models/servico.model';
 import { Solicitacao } from 'src/models/solicitacao.model';
 import { Usuario } from 'src/models/usuario.model';
 import { Veiculo } from 'src/models/veiculo.model';
 import { NotificacaoModule } from '../notificacao/notificacao.module';
 import { SolicitacaoController } from './solicitacao.controller';
-import { SolicitacaoService } from './solicitacao.service';
 
 @Module({
   imports: [
