+++ src/modules/solicitacao/solicitacao.service.ts
@@ -4,6 +4,7 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
+import { UpdateSolicitacaoStatusDto } from './dto/update-solicitacao-status.dto';
 import { Servico } from 'src/models/servico.model';
 import { Solicitacao } from 'src/models/solicitacao.model';
 import { Usuario } from 'src/models/usuario.model';
@@ -73,6 +74,42 @@ export class SolicitacaoService {
 
     return {
       message: 'Solicitação de serviço criada com sucesso',
+    };
+  }
+
+  async findSolicitacaoById(id: number): Promise<Solicitacao> {
+    const solicitacao = await this.solicitacaoModel.findByPk(id);
+
+    if (!solicitacao) {
+      throw new NotFoundException(`Solicitação com ID ${id} não encontrada`);
+    }
+
+    return solicitacao;
+  }
+
+  async updateSolicitacaoStatus(
+    id: number,
+    updateSolicitacaoStatusDto: UpdateSolicitacaoStatusDto,
+  ): Promise<{ message: string }> {
+    const solicitacao = await this.findSolicitacaoById(id);
+
+    const updateData: Partial<Solicitacao> = {
+      status: updateSolicitacaoStatusDto.status,
+    };
+
+    const observacao = updateSolicitacaoStatusDto.observacaoAdmin;
+    if (observacao) {
+      updateData.observacaoAdmin = observacao;
+    }
+
+    if (updateSolicitacaoStatusDto.status === 'concluido') {
+      updateData.dataConclusao = new Date();
+    }
+
+    await solicitacao.update(updateData);
+
+    return {
+      message: 'Status da solicitação atualizado com sucesso.',
     };
   }
 }
